# SAPPORO-service/api-definition/SAPPORO-service-api-definition.yml
swagger: "2.0"
info:
  title: SAPPORO-service REST-API Definition
  description: Specification sheet on SAPPORO-service REST-API
  version: 1.0.0
host: localhost:8002
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /service-info:
    get:
      summary: Get information about Workflow Execution Service.
      description: |-
        May include information related (but not limited to) the workflow descriptor formats, versions supported, the WES API versions supported, and information about general service availability.
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/ServiceInfo"
        "400":
          description: The request is malformed.
          schema:
            $ref: "#/definitions/ErrorResponse"
      tags:
        - WorkflowExecutionService
  /runs:
    get:
      summary: List the workflow runs.
      description: >-
        This list should be provided in a stable ordering.
        (The actual ordering is implementation dependent.)
        When paging through the list, the client should
        not make assumptions about live updates, but should assume the
        contents of the list reflect the workflow list at the moment
        that the first page is requested. To monitor a specific
        workflow run, use GetRunStatus or GetRunLog.
      responses:
        "200":
          description: ""
          type: array
          items:
            $ref: "#/definitions/RunResponse"
        "400":
          description: The request is malformed.
          schema:
            $ref: "#/definitions/ErrorResponse"
      tags:
        - WorkflowExecutionService
    post:
      summary: Run a workflow.
      description: >-
        This endpoint creates a new workflow run and
        returns a `RunId` to monitor its progress.
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/RunResponse"
        "400":
          description: The request is malformed.
          schema:
            $ref: "#/definitions/ErrorResponse"
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: workflow_name
          type: string
        - in: formData
          name: workflow_engine
          type: string
        - in: formData
          name: run_order
          type: file
          format:
            - application/json
            - application/yaml
      tags:
        - WorkflowExecutionService
  /runs/{run_id}:
    get:
      summary: Get detailed info about a workflow run.
      description: >-
        This endpoint provides detailed information about a given workflow run.
        The returned result has information about the outputs produced by this workflow
        (if available), a log object which allows the stderr and stdout to be retrieved,
        a log array so stderr/stdout for individual tasks can be retrieved,
        and the overall state of the workflow run (e.g. RUNNING, see the State section).
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/RunLog"
        "400":
          description: The request is malformed.
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: run_id
          in: path
          required: true
          type: uuid
      tags:
        - WorkflowExecutionService
  /runs/{run_id}/cancel:
    post:
      summary: Cancel a running workflow.
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/RunResponse"
        "400":
          description: The request is malformed.
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: run_id
          in: path
          required: true
          type: uuid
      tags:
        - WorkflowExecutionService
  /workflows:
    get:
      summary: Get workflow list.
      description: >-
        Get workflow list.
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/WorkflowList"
        "400":
          description: The request is malformed.
          schema:
            $ref: "#/definitions/ErrorResponse"
      tags:
        - WorkflowExecutionService
definitions:
  ServiceInfo:
    type: object
    properties:
      supported_wes_versions:
        type: array
        items:
          type: string
          example: v1.0.0
        description: The version(s) of the WES schema supported by this service
      workflow_engines:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
              example: cwltool
            version:
              type: string
              example: 1.0.20181201184214
            workflow_types:
              type: array
              items:
                $ref: "#/definitions/WorkflowType"
            description: >-
              A map with keys as the workflow format type name (currently only CWL and WDL are used
              although a service may support others) and value is a workflow_type_version object which
              simply contains an array of one or more version strings
        description: >-
          The engine(s) used by this WES service, key is engine name (e.g. Cromwell) and value is version
      auth_instructions_url:
        type: string
        example: https://dummy_auth_instructions_url/
        description: >-
          A web page URL with human-readable instructions on how to get an
          authorization token for use with a specific WES endpoint.
      contact_info_url:
        type: string
        example: https://dummy_contact_info_url/
        description: >-
          An email address URL (mailto:) or web page URL with contact information
          for the operator of a specific WES endpoint.  Users of the
          endpoint should use this to report problems or security
          vulnerabilities.
    description: >-
      A message containing useful information about the running service, including supported versions and
      default settings.
  RunResponse:
    type: object
    properties:
      run_id:
        type: uuid
        example: 158c32a8-6baf-4b34-8fdb-8ec5b9c90db7
      status:
        type: string
        example: COMPLETE
        description: The state of the run e.g. (PENDING, RUNNING, FINISH, CANCELED, COMPLETE, SYSTEM_ERROR, EXECUTOR_ERROR)
    description: >-
      A list of workflow runs that the service has executed or is executing.
      The list is filtered to only include runs that the caller has permission to see.
  RunLog:
    type: object
    properties:
      run_id:
        type: uuid
        description: workflow run ID
        example: 158c32a8-6baf-4b34-8fdb-8ec5b9c90db7
      workflow_engine:
        type: string
        example: cwltool
      worklfow_name:
        type: string
        example: bwa_mapping_pe
      run_order:
        type: string
        example: "nthreads: 2 # default value of type"
      status:
        type: string
        example: COMPLETE
        description: The state of the run e.g. (PENDING, RUNNING, FINISH, CANCELED, COMPLETE, SYSTEM_ERROR, EXECUTOR_ERROR)
      stdout:
        type: string
        description: Workflow engine's stdout
      stderr:
        type: string
        description: Workflow engine's stderr
      upload_url:
        type: s3://bucket-name/upload_dir/
  WorkflowType:
    type: object
    properties:
      type:
        type: string
        example: CWL
      version:
        type: string
        example: v1.0
    description: Available workflow types supported by a given instance of the service.
  WorkflowList:
    type: object
    properties:
      workflows:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
              example: bwa_mapping_pe
            type:
              type: string
              example: CWL
            version:
              type: string
              example: v1.0
            content:
              type: string
              description: Workflow file's content.
            run_order_template:
              type: string
              description: Workflow run template.
  ErrorResponse:
    description: >-
      An object that can optionally include information about the error.
    type: object
    properties:
      msg:
        type: string
        description: A detailed error message.
      status_code:
        type: integer
        description: The integer representing the HTTP status code (e.g. 200, 404).